requirements.txt
llama-index
llama-index-llms-openrouter
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-packs-fusion-retriever
sentence-transformers
nest-asyncio
requests
python-dotenv
Flask
Flask-Cors
